# services/api_gateway/routers/telemetry.py
"""
Telemetry Router — Telemetry monitoring, alarms, and metrics endpoints.
"""

from __future__ import annotations

from fastapi import APIRouter, Depends, Query
from typing import Any

from services.api_gateway.auth import User, get_current_user

router = APIRouter()


@router.get("/telemetry/alarms", summary="Get active alarms")
async def get_alarms(
    severity: str | None = Query(None, examples=["critical", "major", "minor"]),
    device_id: str | None = None,
    user: User = Depends(get_current_user),
):
    """List active alarms, optionally filtered by severity or device."""
    return {"alarms": [], "total": 0, "filters": {"severity": severity, "device_id": device_id}}


@router.get("/telemetry/alarms/{alarm_id}", summary="Get alarm details")
async def get_alarm(alarm_id: str, user: User = Depends(get_current_user)):
    return {"alarm_id": alarm_id, "status": "unknown"}


@router.post("/telemetry/alarms/{alarm_id}/acknowledge", summary="Acknowledge an alarm")
async def acknowledge_alarm(alarm_id: str, user: User = Depends(get_current_user)):
    return {"alarm_id": alarm_id, "acknowledged": True}


@router.get("/telemetry/metrics", summary="Query telemetry metrics")
async def get_metrics(
    device_id: str | None = None,
    metric_path: str | None = None,
    limit: int = Query(100, ge=1, le=1000),
    user: User = Depends(get_current_user),
):
    """Query telemetry metrics with optional filtering."""
    return {"metrics": [], "total": 0}


@router.get("/telemetry/correlations", summary="Get 1-3-5 correlation sessions")
async def get_correlations(user: User = Depends(get_current_user)):
    """List active and completed 1-3-5 troubleshooting correlation sessions."""
    return {"correlations": [], "total": 0}


@router.get("/telemetry/remediations", summary="Get remediation history")
async def get_remediations(user: User = Depends(get_current_user)):
    """List autonomous remediation actions taken by the 1-3-5 framework."""
    return {"remediations": [], "total": 0}